backend/requirements.txt
openai
python-dotenv
jira
watchdog
flask
flask-cors
requests
